Rework that data flow
