+++ src/applications/coronavirus-vaccination/routes.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 
-import CoronavirusVaccinationApp from './containers/CoronavirusVaccinationApp';
+import Layout from './components/Layout';
 import Introduction from './components/Introduction';
 import Form from './components/Form';
 import Confirmation from './components/Confirmation';
 
 const routes = (
   <Route path="/">
-    <Route component={CoronavirusVaccinationApp} key="/main">
+    <Route component={Layout} key="/main">
       <IndexRoute component={Introduction} key="/intro" />
       <Route component={Form} key="/apply" path="/apply" />
       <Route
